% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicShiftColR.R
\name{cyclicShiftColR}
\alias{cyclicShiftColR}
\title{Perform the cyclic shift procedure on the columns of a matrix}
\usage{
cyclicShiftColR(X, randomSeed = NULL)
}
\arguments{
\item{X}{a matrix or a data frame of copy number data. The rows and columns
of X correspond to genes and subjects, respectively.}

\item{randomSeed}{a random seed.  Default = NULL.}
}
\value{
a matrix Z whose dimensions are the same as X.  Each column of Z
is obtained by perform a cyclic shift of the corresponding column of X.
}
\description{
Perform the cyclic shift procedure on the columns of a matrix
}
\examples{
test = matrix(c(1:50), 10, 5)
cyclicShiftColR(test, randomSeed = NULL)

}
