% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{Prepare copy number data for downstream analysis}
\usage{
dataPrep(X, Y = NULL, species = c("human", "mouse"))
}
\arguments{
\item{X}{a matrix or a data frame of copy number data. The rows and columns
of X correspond to genes and subjects, respectively. X must have gene
symbols as its row names.}

\item{Y}{an optional matrix or data frame of copy number data to compare with
X. As X, the rows and columns of Y correspond to genes and subjects,
respectively. Y must have gene symbols as its row names.}

\item{species}{a value specifying species for ensembl database to be used;
Default is "human".}
}
\value{
a list of processed X and Y with a data frame containing gene
annotation information.
}
\description{
Prepare copy number data for downstream analysis
}
\examples{
\donttest{
#This runtime for this code slightly exceeds the limits imposed by CRAN.
data(DiNAMIC.Duo)
output = dataPrep(X=luadSubset,Y=NULL)
}

}
