% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeChrPlot.R
\name{genomeChrPlot}
\alias{genomeChrPlot}
\title{A Function for Plotting Mean Copy Number Values and Differences Across Multiple Chromosomes}
\usage{
genomeChrPlot(
  inputList,
  plottingChrs = NULL,
  lwdVec = rep(1, 3),
  ltyVec = c(1:3),
  lineColorVec = c("red", "blue", "black"),
  ylimLow = -1,
  ylimHigh = 1,
  chrLabel = TRUE,
  xaxisLabel = "Chromosome",
  yaxisLabel = NULL,
  mainLabel = NULL,
  axisCex = 1,
  labelCex = 1,
  xaxisLine = 2.5,
  yaxisLine = 2.5,
  mainLine = 0,
  marginVec = c(4, 4, 3, 3),
  legendText = NULL,
  highThreshold = NULL,
  lowThreshold = NULL,
  showLegend = FALSE,
  legendXQuantile = 0.55,
  legendYCoord = 1
)
}
\arguments{
\item{inputList}{A list produced by dataPrep.}

\item{plottingChrs}{A numeric list of chromosomes to be plotted.  One plot for each chromosome.}

\item{lwdVec}{A vector of line widths.  Default = rep(1, 3).  See \code{\link{par}}.}

\item{ltyVec}{A vector of line types.  Default = c(1:3).  See \code{\link{par}}.}

\item{lineColorVec}{A vector of line colors.  Default = c("red", "blue", "black").}

\item{ylimLow}{The lower limit of the y-values in the plot.  Default = -1.  See \code{\link{plot}}.}

\item{ylimHigh}{The upper limit of the y-values in the plot.  Default = 1.  See \code{\link{plot}}.}

\item{chrLabel}{Binary value determining whether or not chromosomes are labeled.  Default = TRUE.}

\item{xaxisLabel}{Label for the x-axis in the plot.  Default = "Chromosome".  See \code{\link{plot}}.}

\item{yaxisLabel}{Label for the y-axis in the plot.  Default = NULL.  See \code{\link{plot}}.}

\item{mainLabel}{Main label in the plot.  Default = NULL.  See \code{\link{plot}}.}

\item{axisCex}{Point size for the scale on the axis.  Default = 1.  See \code{\link{par}}.}

\item{labelCex}{Point size for the axis label.  Default = 1.  See \code{\link{par}}.}

\item{xaxisLine}{Numerical value used to specify the location of the x-axis label.  Default = 2.5.  See \code{\link{mtext}}.}

\item{yaxisLine}{Numerical value used to specify the location of the y-axis label.  Default = 2.5.  See \code{\link{mtext}}.}

\item{mainLine}{Numerical value used to specify the location of the main.label.  Default = 0.  See \code{\link{mtext}}.}

\item{marginVec}{Numerical vector specifying margin sizes.  Default = c(4, 4, 3, 3).  See \code{\link{par}}.}

\item{legendText}{Character vector used to legend.  Only shown if showLegend = TRUE.  Default = NULL.  See \code{\link{legend}}.}

\item{highThreshold}{Numerical value representing the position of the upper horizontal line.  Default = NULL.}

\item{lowThreshold}{Numerical value representing the position of the lower horizontal line.  Default = NULL.}

\item{showLegend}{Binary value determining whether or not the legend is shown.  Default = FALSE.  See \code{\link{legend}}.}

\item{legendXQuantile}{Quantile to specify the "x" location of the legend.  Only relevant if showLegend = TRUE  Default = 0.55.  See \code{\link{legend}}.}

\item{legendYCoord}{Numerical value to specify the "y"location of the legend.  Only relevant if showLegend = TRUE.  Default = 1.  See \code{\link{legend}}.}
}
\value{
Creates a multi-page plot of mean copy number values and differences by chromosome.
}
\description{
This function plots mean copy number values from one or two cohorts at a common set of markers across multiple chromosomes.
}
\examples{
genomeChrPlot(inputList = pD, ylimLow = -1.4, ylimHigh = 1.4)

}
