% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luadSubset.R
\docType{data}
\name{luadSubset}
\alias{luadSubset}
\title{DNA copy number data for lung adenocarcinoma.}
\format{
A numeric matrix with 3475 rows and 65 columns:
\describe{
\item{}{Genes appear in rows}
\item{}{Samples appear in columns}
}
}
\source{
\url{https://gdac.broadinstitute.org/}
}
\usage{
luadSubset
}
\description{
A subset of the DNA copy number data from the TCGA lung
lung adenocarcinoma cohort.
}
\keyword{datasets}
