% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luscSubset.R
\docType{data}
\name{luscSubset}
\alias{luscSubset}
\title{DNA copy number data for lung squamous cell carcinoma.}
\format{
A numeric matrix with 3480 rows and 60 columns:
\describe{
\item{}{Genes appear in rows}
\item{}{Samples appear in columns}
}
}
\source{
\url{https://gdac.broadinstitute.org/}
}
\usage{
luscSubset
}
\description{
A subset of the DNA copy number data from the TCGA lung
lung squamous cell carcinoma cohort.
}
\keyword{datasets}
