% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\docType{data}
\name{miniconda_installation}
\alias{miniconda_installation}
\title{Perform necessary tasks when the DiNAMICDuo package is loaded}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
miniconda_installation
}
\description{
Perform necessary tasks when the DiNAMICDuo package is loaded
}
\keyword{datasets}
