% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pD.R
\docType{data}
\name{pD}
\alias{pD}
\title{Prepped data for DiNAMIC.Duo.}
\format{
A list containing three components:
\describe{
\item{}{DNA copy number data for lung adenocarcinoma}
\item{}{DNA copy number data for lung squamous cell carcinoma}
\item{}{Gene position data}
\item{}{These components are available for a common set of genes}
}
}
\usage{
pD
}
\description{
A list produced by the dataPrep() function.
}
\keyword{datasets}
