% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsProcess.R
\name{resultsProcess}
\alias{resultsProcess}
\title{Processing peeling results}
\usage{
resultsProcess(peel.results, posDT)
}
\arguments{
\item{peel.results}{peeling results}

\item{posDT}{a data frame containing gene annotation information; a list
component created by \code{dataPrep}.}
}
\value{
processed peeling results with a list of genes corresponding to each
peeled region
}
\description{
Processing peeling results
}
\seealso{
\code{\link{dataPrep}}
}
